cookbook/mcp_servers/exchange_rate/pyproject.toml
[project]
name = "exchange-rate-mcp-server"
version = "0.1.0"
description = "An exchange rate MCP server for agentstr."
dependencies = [
    "python-dotenv",
    "httpx",
    "agentstr @ {root:uri}/../../..",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
